+++ packages/util/src/object/copy.ts
@@ -3,6 +3,10 @@
 
 import { objectSpread } from './spread';
 
+/**
+ * @name objectCopy
+ * @summary Creates a shallow clone of the input object
+ */
 export function objectCopy <T extends object> (source: T): T {
   return objectSpread({}, source);
 }