+++ packages/util/src/object/spread.ts
@@ -1,6 +1,10 @@
 // Copyright 2017-2021 @polkadot/util authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+/**
+ * @name objectSpread
+ * @summary Concats all sources into the destination
+ */
 export function objectSpread <T extends object> (dest: object, ...sources: (object | undefined | null)[]): T {
   for (let i = 0; i < sources.length; i++) {
     const src = sources[i];